requirements.txt
openai
pydantic
pyyaml
openenv-core
fastapi
uvicorn
websockets
